+++ sugar/dsn.go
@@ -2,6 +2,8 @@ package sugar
 
 import "net/url"
 
+type dsnOption func(dsn *url.URL)
+
 // Usage of this package
 //
 // db, err := ydb.Open(ctx,
@@ -9,7 +11,7 @@ import "net/url"
 // )
 
 // DSN makes connection string (data source name) by endpoint, database and secure
-func DSN(endpoint, database string, secure bool) (s string) {
+func DSN(endpoint, database string, opts ...dsnOption) (s string) {
 	qp := url.Values{}
 
 	dsn := url.URL{
@@ -19,9 +21,25 @@ func DSN(endpoint, database string, secure bool) (s string) {
 		RawQuery: qp.Encode(),
 	}
 
-	if secure {
-		dsn.Scheme = "grpcs"
+	for _, opt := range opts {
+		opt(&dsn)
 	}
 
 	return dsn.String()
+}
+
+func WithSecure(secure bool) dsnOption {
+	return func(dsn *url.URL) {
+		if secure {
+			dsn.Scheme = "grpcs"
+		} else {
+			dsn.Scheme = "grpc"
+		}
+	}
+}
+
+func WithUserPassword(user string, password string) dsnOption {
+	return func(dsn *url.URL) {
+		dsn.User = url.UserPassword(user, password)
+	}
 }