+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+* Added `sugar.WithUserPassword(user,password)` option for `sugar.DSN()` helper
+* Added `sugar.WithSecure(bool)` option for `sugar.DSN()` helper
+* Small breaking change: `sugar.DSN` have only two required parameters (endpoint and database) from now on. 
+  Third parameter `secure` must be passed as option `sugar.WithSecure(bool)`
+
 ## v3.92.0
 * Added experimental ydb.ParamsFromMap and ydb.MustParamsFromMap for build query parameters
 * Refactored coordination traces